server/requirements.txt
black==24.10.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mypy-extensions==1.0.0
numpy==2.2.1
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
threadpoolctl==3.5.0
urllib3==2.3.0
Werkzeug==3.1.3